django-template-engine/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-crispy-forms==1.11.2
django-sslserver==0.22
Django-Verify-Email==1.0.7
idna==2.10
oauthlib==3.1.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
